api: move types.Volume to volume.Volume
(cherry picked from commit f19ef20a4414a6f63da8fb2493e6010359652ad1)
