+++ hack/generate-swagger-api.sh
@@ -8,10 +8,13 @@ swagger generate model -f api/swagger.yaml \
 	-n IdResponse \
 	-n ImageDeleteResponseItem \
 	-n ImageSummary \
-	-n Plugin -n PluginDevice -n PluginMount -n PluginEnv -n PluginInterfaceType \
+	-n Plugin \
+	-n PluginDevice \
+	-n PluginMount \
+	-n PluginEnv \
+	-n PluginInterfaceType \
 	-n Port \
-	-n ServiceUpdateResponse \
-	-n Volume
+	-n ServiceUpdateResponse
 
 swagger generate model -f api/swagger.yaml \
 	-t api -m types/container --skip-validator -C api/swagger-gen.yaml \
@@ -20,6 +23,7 @@ swagger generate model -f api/swagger.yaml \
 
 swagger generate model -f api/swagger.yaml \
 	-t api -m types/volume --skip-validator -C api/swagger-gen.yaml \
+	-n Volume \
 	-n VolumeCreateOptions
 
 swagger generate operation -f api/swagger.yaml \
